pyproject.toml
[project]
name = "gen_pixel_art"
version = "0.1.0"
description = "Converts text-to-art generated 'pixel' art from LLMs to true resolution pixel art assets."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.4",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
]
packages = [
    { include = "gen_pixel_art" }
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pylint>=3.3.7",
]
